requirements.txt
pandas
streamlit
st-gsheets-connection 
gspread
google-genai
python-dotenv